+++ apps/web/src/app/_pages/login/_common/form-login/form-login.ts
@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import { Component, signal } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormField } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
@@ -10,4 +11,5 @@ import { RouterLink } from '@angular/router';
   templateUrl: './form-login.html',
   styleUrl: './form-login.css',
 })
-export class FormLogin {}
+export class FormLogin {
+}