+++ apps/web/src/app/_service/categoria-tarefa-service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CategoriaTarefaService } from './categoria-tarefa-service';
+
+describe('CategoriaTarefaService', () => {
+  let service: CategoriaTarefaService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CategoriaTarefaService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});