+++ apps/web/src/app/_service/categoria-tarefa-service.ts
@@ -1,0 +1,11 @@
+import { HttpClient } from '@angular/common/http';
+import { Service } from '@angular/core';
+
+@Service()
+export class CategoriaTarefaService {
+    constructor(private http: HttpClient) {}
+
+    findAll() {
+        return this.http.get("");
+    }
+}